requirements.txt
jupyterlab
python-dotenv
langchain
openai
